fixing backend

+++ MealsManagement/app_two/views.py
@@ -11,6 +11,7 @@ def startvote(request):
     if 'startvote' in request.session:
         return redirect('/voting')
     else:
+        request.session['startvote'] = 1
         context={
             'show_all_resturants': models.show_all_resturants(),
             "logged_user" : models.get_specific_user(request),
@@ -33,31 +34,32 @@ def submitvote(request,ic):
 
 def create_msg(request):
     models.create_message(request)
-    return redirect('/')
+    return redirect('/voting')
 
 
 def thewinner(request):
-    del request.session['startvote']
-    # models.Restaurant.through.objects.all().delete()
-    models.Restaurant.users_who_voted.through.objects.all().delete()
-    
-    rest = models.Restaurant.objects.all()
-    for res in rest : 
-        res.votes=0
-        res.save()
+    if 'startvote' not in request.session:
+        return redirect('/startvote')
+    else:
+        del request.session['startvote']
+        models.Restaurant.users_who_voted.through.objects.all().delete()
+        msg=models.Message.objects.all()
+        rest = models.Restaurant.objects.all()
+        for mseg in msg : 
+            mseg.delete()
+        for res in rest : 
+            res.votes=0
+            res.save()
     # b.users_who_voted.aclear()
-    context={
-        'the_winner_rest': models.get_the_winner_rest(request),
-        "logged_user" : models.get_specific_user(request),
-    }
-    return render(request,'thewinner.html',context)
+        context={
+            'the_winner_rest': models.get_the_winner_rest(request)[0],
+            "logged_user" : models.get_specific_user(request),
+        }
+        return render(request,'thewinner.html',context)
 
-def testwinner(request):    
-    context={
-        'the_winner_rest': models.get_the_winner_rest(),
-        "logged_user" : models.get_specific_user(request),
-    }
-    return render(request,'thewinner.html',context)
+
+
+
 
 def add_a_company(request):
     return render (request,'addcompanypage.html')
